+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -6,6 +6,10 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import roomescape.domain.Reservation;
+import roomescape.dao.ReservationQueryingDAO;
+import roomescape.dao.ReservationUpdatingDAO;
+import roomescape.domain.Time;
 
 @RestController
 public class ReservationController {
