+++ src/main/java/roomescape/dao/ReservationQueryingDAO.java
@@ -1,10 +1,12 @@
-package roomescape;
+package roomescape.dao;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
+import roomescape.domain.Reservation;
+import roomescape.domain.Time;
 
 @Repository
 public class ReservationQueryingDAO {
