+++ src/main/java/roomescape/dao/ReservationUpdatingDAO.java
@@ -1,14 +1,13 @@
-package roomescape;
+package roomescape.dao;
 
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 
 import java.util.HashMap;
 import java.util.Map;
+import roomescape.domain.Reservation;
+import roomescape.exception.GlobalExceptionHandler;
 
 @Repository
 public class ReservationUpdatingDAO {
@@ -34,15 +33,11 @@ public class ReservationUpdatingDAO {
 
     public int delete(long id){
         String sql = "DELETE FROM reservation WHERE id = ?";
+
         int num = jdbcTemplate.update(sql, id);
         if(num == 0){
             throw new GlobalExceptionHandler.NotFoundReservationException("Reservation is not found with id: " + id);
         }
         return num;
-    }
-
-    @ExceptionHandler(GlobalExceptionHandler.NotFoundReservationException.class)
-    public ResponseEntity handleNotFoundReservationException(GlobalExceptionHandler.NotFoundReservationException e) {
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
     }
 }
