+++ src/main/java/roomescape/domain/Reservation.java
@@ -1,11 +1,9 @@
-package roomescape;
+package roomescape.domain;
 
 //import jakarta.persistence.Entity;
 //import jakarta.persistence.GeneratedValue;
 //import jakarta.persistence.GenerationType;
 //import jakarta.persistence.Id;
-
-import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 import java.time.LocalDate;
