+++ src/main/java/roomescape/exception/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
