+++ src/test/java/roomescape/ReservationDaoTest.java
@@ -1,13 +1,13 @@
 package roomescape;
 
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 
-import javax.naming.spi.ResolveResult;
 import java.util.List;
+import roomescape.dao.ReservationQueryingDAO;
+import roomescape.domain.Reservation;
 
 import static org.assertj.core.api.Assertions.assertThat;
